EasyBin compiles to lib,  this is a stopgap
the plan is to get code to sort of compile by not compiling files that
have a hard dependency on files we don't have.  this will mean that main
functions won't link, and we will likely try and reference unimplemented
functions.  As a stopgap, we can convert executable targets to static
libs, since they compile code, but don't care if code links or not